requirements.txt
# Core dependencies
httpx>=0.25.0
pyyaml>=6.0
jinja2>=3.1.0
numpy>=1.24.0
python-dotenv>=1.0.0
rich>=13.0.0

# Web server
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Semantic similarity (optional, comment out if not needed)
sentence-transformers>=2.2.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Report generation
pandas>=2.0.0
